requirements.txt
networkx
distro
future
pygame
numpy
pydantic
open3d-python
pytest
typer
typer[all]
PyYAML
carla~=0.9.15